refactored drift methods.
