+++ setup.py
@@ -1,10 +1,10 @@
 from setuptools import setup, find_packages
 
 setup(name='GSadjust',
-      version='0.0.1',
+      version='0.98',
       description='Gravity-survey network adjustment',
       author='',
-      author_email='',
-      url='',
+      author_email='jkennedy@usgs.gov',
+      url='https://github.com/jkennedy-usgs/sgp-gsadjust',
       packages=find_packages(exclude=['test_data', 'docs', 'tests*']),
      )
